All deployer tests ok

+++ tools/custom_fs_classes/DFile.class.php
@@ -107,7 +107,10 @@ class DFile extends DFileSystemElement
  */
     function delete()
     {
-        return @unlink($this->__full_path);
+        if (file_exists($this->__full_path))
+            return @unlink($this->__full_path);
+        else 
+            return false;
     }
 
     function isEmpty()
